third_party/blink/web_tests/wpt_internal/prerender/restriction-web-xr-immersive-vr-session.https.html

<!DOCTYPE html>
<title>Access to the WebXR immersive-vr session API is deferred</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('test-channel');
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  const url = `resources/web-xr-immersive-vr-session.https.html`;
  window.open(url, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {
      event: `started waiting XRSession.requestSession('immersive-vr')`,
      prerendering: true
    },
    {
      event: `prerendering change`,
      prerendering: false
    },
    {
      event: `XRSession.requestSession('immersive-vr') rejected: SecurityError`,
      prerendering: false
    },
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event[${i}]`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering[${i}]`);
  }
}, `The access to the WebXR immersive-vr session API should be deferred until
    the prerendered page is activated`);

</script>
</body>
